pyproject.toml
[project]
name = "s8mcpgoestointernet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.169.0",
    "google-auth>=2.39.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.13.0",
    "mcp>=1.7.0",
    "pytelegrambotapi>=4.26.0",
    "python-dotenv>=1.1.0",
    "asyncio>=3.4.3",
    "bs4>=0.0.2",
    "faiss-cpu>=1.10.0",
    "httpx>=0.28.1",
    "llama-index>=0.12.31",
    "llama-index-embeddings-google-genai>=0.1.0",
    "markitdown[all]>=0.1.1",
    "mcp[cli]>=1.6.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pymupdf4llm>=0.0.21",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
    "trafilatura[all]>=2.0.0",
    "fastapi>=0.115.12",
]
